fix: Deleted asso roles classes and integrated them in user and associations classes

+++ bot/bdd/classes/associationsManager.js
@@ -1,9 +1,11 @@
-const {Association} = require('./association.js');
+const { Association } = require('./association.js');
 
 class AssociationsManager
 {
-    static async getAssociations(name)
+    static async getAssociation(name)
     {
+        name = name.toLowerCase();
+        name = name.replace(/ISEN/g, "").replace(/ /g, "");
         const connection = global.sqlConnection;
         const query = "SELECT * FROM association WHERE name = ?";
         const data = await connection(query, name);
@@ -12,31 +14,49 @@ class AssociationsManager
             return new Association(data.asso_id);
     }
 
-    static async getAssociationsByID(asso_id)
+    static async getAssociationByID(asso_id)
     {
         const connection = global.sqlConnection;
         const query = "SELECT * FROM association WHERE asso_id = ?";
-        const data = await connection(query, asso_id);
+        const data = await connection(query, [asso_id]);
         if (data.length === 0) return null;
         else
             return new Association(asso_id);
     }
 
+    static async getAllAssociations()
+    {
+        const connection = global.sqlConnection;
+        const query = "SELECT * FROM association";
+        const data = await connection(query);
+        if (data.length === 0) return null;
+        else
+        {
+            let associations = [];
+            for (let i = 0; i < data.length; i++)
+                associations.push(await this.getAssociationByID(data[i].asso_id));
+            return associations;
+        }
+    }
+
     static async addAssociation(association)
     {
+        if (await this.getAssociation(association.name) !== null) return;
         const connection = global.sqlConnection;
         const query = "INSERT INTO association (name, description, icon) VALUES (?, ?, ?)";
         const values = [association.name, association.description, association.icon];
         await connection(query, values);
-        return await this.getAssociations(association.name);
+        return await this.getAssociation(association.name);
     }
 
     static async deleteAssociation(asso_id)
     {
+        if (await this.getAssociationByID(asso_id) === null) return;
         const connection = global.sqlConnection;
         const query = "DELETE FROM association WHERE asso_id = ?";
         await connection(query, asso_id);
     }
+
 }
 
 module.exports = { AssociationsManager };
